backend/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
grpcio-reflection==1.60.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0
